chore: storybookにfont-family反映、background color設定
